+++ src/main/java/com/bookbla/americano/domain/memberask/repository/entity/MemberAsk.java
@@ -1,4 +1,4 @@
-package com.bookbla.americano.domain.memberask;
+package com.bookbla.americano.domain.memberask.repository.entity;
 
 import com.bookbla.americano.base.entity.BaseInsertEntity;
 import com.bookbla.americano.domain.member.Member;
