+++ src/main/java/com/bookbla/americano/domain/postcard/Postcard.java
@@ -2,7 +2,7 @@ package com.bookbla.americano.domain.postcard;
 
 import com.bookbla.americano.base.entity.BaseInsertEntity;
 import com.bookbla.americano.domain.member.Member;
-import com.bookbla.americano.domain.memberask.MemberReply;
+import com.bookbla.americano.domain.memberask.repository.entity.MemberReply;
 import com.bookbla.americano.domain.quiz.QuizReply;
 import javax.persistence.Entity;
 import javax.persistence.EnumType;
